[refactor] 가격 Double 에서 Integer 로 변경
issue: 55

+++ src/main/java/com/mini/joymall/order/domain/entity/OrderItem.java
@@ -24,16 +24,16 @@ public class OrderItem {
     private Long productOptionId;
 
     private Integer quantity;
-    private Double pricePerItem;
+    private Integer pricePerItem;
     private LocalDateTime createdDate;
     private LocalDateTime updatedDate;
 
-    public OrderItem(Long productId, Long productOptionId, Integer quantity, Double pricePerItem) {
+    public OrderItem(Long productId, Long productOptionId, Integer quantity, Integer pricePerItem) {
         this(productId, productOptionId, quantity, pricePerItem, LocalDateTime.now(), LocalDateTime.now());
     }
 
     @Builder
-    public OrderItem(Long productId, Long productOptionId, Integer quantity, Double pricePerItem, LocalDateTime createdDate, LocalDateTime updatedDate) {
+    public OrderItem(Long productId, Long productOptionId, Integer quantity, Integer pricePerItem, LocalDateTime createdDate, LocalDateTime updatedDate) {
         this.productId = productId;
         this.productOptionId = productOptionId;
         this.quantity = quantity;
@@ -42,7 +42,7 @@ public class OrderItem {
         this.updatedDate = updatedDate;
     }
 
-    public Double calculateTotalPrice() {
+    public Integer calculateTotalPrice() {
         return quantity * pricePerItem;
     }
 }
